+++ apiTools/v1/changelog.md
@@ -1,8 +1,13 @@
 # Changelog
 
+## [0.4.1-alpha] - 2023-05-21
+- Nuevo endpoint getAdminGroupById
+- Nuevo endpoint putAdminGroups
+
 ## [0.4.0-alpha] - 2023-05-21
 - Tareas
 - Horarios
 - Alertas
 - Grupos
+