+++ changelog.md
@@ -1,7 +1,12 @@
 # Changelog
 
-## [0.10.0-alpha] - 2023-05-21
-- ApiUsers
+## [0.10.3-alpha] - 2023-05-21
 - ApiTools
-- ApiRepos
+
+## [0.10.2-alpha] - 2023-05-21
+- ApiUsers
+
+
+
+